Add data sort capability to the coin table
- By clicking on a table column heading the table data is sorted (based upon the clicked column's data) into either ascending or descending order.

- A second click on a same column header will reverse the sort order (ie. ascending data will be reversed to show it in descending order, and vica versa).

- The sorting capability may be applied to the coin table columns containing either numeric or string data.  It has so far been applied to the columns: Rank, Coin (name), Price, Change (24hr), and Market Cap.

Note: To add the same sorting functionality to further coin table columns:
  1. Add the event handler to relevant column header (ie. in html <th> tag): onclick="sortTableRows(event)"
  2. Add that column's coin data attribute into the switch statement, within function getCoinObjectAttribute(IdElement); that is in logic.js

Note: The functionality sorts previously obtained coin data (held as an array of objects) that is already displayed in the table (ie. it doesn't obtain the latest/updated data via a new api call but rather sorts the currently held data).

Note: Styling of the column headers, to show columns that can be clicked and/ or if data has been sorted, has yet to be implemented.

-